backend/requirements-fastapi.txt
# ============================================
# CORE FASTAPI
# ============================================
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
pydantic-settings==2.6.0
python-multipart==0.0.20

# ============================================
# DATABASE (PostgreSQL Async)
# ============================================
sqlalchemy==2.0.36
alembic==1.14.0
psycopg[binary]>=3.2.3  # ✅ Substituimos asyncpg + psycopg2-binary
psycopg-pool>=3.2.0

# ============================================
# AUTHENTICATION & SECURITY
# ============================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# ============================================
# ASYNC SUPPORT
# ============================================
aiofiles==24.1.0
python-dotenv==1.0.1
aiosmtplib==3.0.1

# ============================================
# RATE LIMITING
# ============================================
slowapi==0.1.9

# ============================================
# COMPUTER VISION (MANTER DO PROJETO ATUAL)
# ============================================
opencv-python==4.10.0.84
ultralytics==8.3.65
numpy==2.2.6
shapely==2.0.6
psutil==6.1.1
pillow==12.0.0
